move the rest of databricks files
