+++ sql-plugin/src/main/301db/scala/org/apache/spark/sql/rapids/shims/v2/GpuFileScanRDD.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021, NVIDIA CORPORATION.
+ * Copyright (c) 2020-2021, NVIDIA CORPORATION.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.spark.sql.rapids.shims.spark311db
+package org.apache.spark.sql.rapids.shims.v2
 
 import java.io.{FileNotFoundException, IOException}
